Rename index.tsx to interactive-graph-editor.tsx
