+++ packages/perseus-editor/src/widgets/__stories__/interactive-graph-editor.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import {flags} from "../../__stories__/flags-for-api-options";
 import {getDefaultFigureForType} from "../../components/util";
-import InteractiveGraphEditor from "../interactive-graph-editor";
+import InteractiveGraphEditor from "../interactive-graph-editor/interactive-graph-editor";
 
 import InteractiveGraphEditorArgTypes from "./interactive-graph-editor.argtypes";
