+++ packages/perseus-editor/src/widgets/__tests__/interactive-graph-editor-locked-figures.test.tsx
@@ -7,7 +7,7 @@ import * as React from "react";
 import {testDependencies} from "../../../../../testing/test-dependencies";
 import {flags} from "../../__stories__/flags-for-api-options";
 import {getDefaultFigureForType} from "../../components/util";
-import InteractiveGraphEditor from "../interactive-graph-editor";
+import InteractiveGraphEditor from "../interactive-graph-editor/interactive-graph-editor";
 
 import type {PerseusGraphType} from "@khanacademy/perseus";
 import type {UserEvent} from "@testing-library/user-event";
